+++ Sources/ViFlow/macOS/ViFlowUpdate/AppDelegate.h
@@ -1,0 +1,13 @@
+//
+//  AppDelegate.h
+//  ViFlowUpdate
+//
+//  Created by Do Vinh Quang on 5/15/26.
+//  Copyright © 2026 Do Vinh Quang. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface AppDelegate : NSObject <NSApplicationDelegate>
+
+@end